added growth dto and domain model
