+++ app/src/main/java/com/eidorian/weedtacora/data/repository/BinnacleRepository.kt
@@ -1,10 +1,10 @@
 package com.eidorian.weedtacora.data.repository
 
+import com.eidorian.weedtacora.data.dto.GrowthWithBinnacle
 import com.eidorian.weedtacora.data.entities.Binnacle
-import com.eidorian.weedtacora.data.entities.GrowthWithBinnacle
 
 interface BinnacleRepository {
     suspend fun insertBinnacle(binnacle: Binnacle): Long
-    suspend fun getAllBinnacles(): List<GrowthWithBinnacle>
+    //suspend fun getAllBinnacles(): List<GrowthWithBinnacle>
     suspend fun getGrowthDetails(growthId: Int): List<Binnacle>
 }
